+++ src/main/java/springrestapidemo/service/UserService.java
@@ -1,11 +1,15 @@
 package springrestapidemo.service;
 
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import springrestapidemo.dto.UserDto;
 import springrestapidemo.entity.UserEntity;
 import springrestapidemo.repository.UserRepository;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 /**
  * @author Nikita Gvardeev
@@ -13,32 +17,60 @@ import java.util.List;
  */
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class UserService {
 
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
 
-    public UserEntity findById(Long id) {
-        return userRepository.findById(id).get();
+    public UserDto findById(Long id) {
+        UserEntity userEntity = userRepository.findById(id).orElse(null);
+
+        if (Objects.isNull(userEntity))
+            throw new UsernameNotFoundException("User by id: " + id + " not found");
+
+        return UserDto.toDto(userEntity);
     }
 
-    public UserEntity findByEmail(String email) {
-        return userRepository.findByEmail(email);
+    public UserDto findByEmail(String email) {
+        UserEntity userEntity = userRepository.findByEmail(email);
+
+        if (Objects.isNull(userEntity))
+            throw new UsernameNotFoundException("User by email: " + email + " not found");
+
+        return UserDto.toDto(userEntity);
     }
 
-    public List<UserEntity> findAll() {
-        return userRepository.findAll();
+    public List<UserDto> findAll() {
+        return userRepository.findAll()
+                .stream()
+                .map(UserDto::toDto)
+                .collect(Collectors.toList());
     }
 
-    public UserEntity save(UserEntity userEntity) {
-        return userRepository.save(userEntity);
+    public UserDto save(UserDto userDto) {
+        return UserDto
+                .toDto(userRepository
+                        .save(UserDto
+                                .toEntity(userDto)));
     }
 
-    public UserEntity update(UserEntity userEntity) {
-        return userRepository.save(userEntity);
+    public UserDto update(UserDto userDto, Long id) {
+        UserEntity userEntity = userRepository.findById(id).orElse(null);
+
+        if (Objects.isNull(userEntity))
+            throw new UsernameNotFoundException("User by id: " + id + " not found");
+
+        userEntity.setEmail(userDto.getEmail());
+
+        return UserDto.toDto(userRepository.save(userEntity));
     }
 
-    public void delete(UserEntity userEntity) {
+    public void delete(Long id) {
+        UserEntity userEntity = userRepository.findById(id).orElse(null);
+
+        if (Objects.isNull(userEntity))
+            throw new UsernameNotFoundException("User by id: " + id + " not found");
+
         userRepository.delete(userEntity);
     }
 }
